1、代码优化 2、bug修复 3、数据库优化

+++ crmeb/src/main/java/com/zbkj/crmeb/front/response/PriceGroupResponse.java
@@ -22,5 +22,5 @@ public class PriceGroupResponse {
     private BigDecimal payPrice;
     private BigDecimal payPostage;
     private BigDecimal deductionPrice;
-    private BigDecimal usedIntegral;
+    private Integer usedIntegral;
 }
